Only clean courses on sundays

+++ app/workers/clean_courses_scheduler.rb
@@ -1,9 +1,12 @@
 class CleanCoursesScheduler
 
   def perform
-    date = Date.today
-    date = date - date.wday + 1
-    Course.where('date < ?', date).destroy_all
+    # Only clean courses on sundays
+    if Date.today.wday == 0
+      date = Date.today
+      date = date - date.wday + 1
+      Course.where('date < ?', date).destroy_all
+    end
   end
 
 end
